Group docker compose wrapper under service cmd
This separation of concern has the benefit of being able
to have each wrapper as its own plugin if needed.
