+++ src/commands/service/logs.ts
@@ -1,8 +1,8 @@
 import {flags} from '@oclif/command'
 
-import {dryRunFlag} from '../flags'
-import DockerComposeCommand from '../wrapper/docker-compose'
-import {environmentFlag, servicesFlag} from '../wrapper/docker-compose/flags'
+import {dryRunFlag} from '../../flags'
+import DockerComposeCommand from '../../wrapper/docker-compose'
+import {environmentFlag, servicesFlag} from '../../wrapper/docker-compose/flags'
 
 export default class Logs extends DockerComposeCommand {
   static description = 'show service logs'
