+++ src/commands/service/start.ts
@@ -1,11 +1,11 @@
 import {flags} from '@oclif/command'
 
-import {dryRunFlag} from '../flags'
-import DockerComposeCommand from '../wrapper/docker-compose'
-import {environmentFlag, servicesFlag} from '../wrapper/docker-compose/flags'
+import {dryRunFlag} from '../../flags'
+import DockerComposeCommand from '../../wrapper/docker-compose'
+import {environmentFlag, servicesFlag} from '../../wrapper/docker-compose/flags'
 
-export default class ReStart extends DockerComposeCommand {
-  static description = 'stop, (re)create and start services in daemon mode'
+export default class Start extends DockerComposeCommand {
+  static description = '(re)create and start services in daemon mode'
 
   static flags = {
     services: servicesFlag,
@@ -15,7 +15,7 @@ export default class ReStart extends DockerComposeCommand {
   }
 
   async run() {
-    const {flags} = this.parse(ReStart)
+    const {flags} = this.parse(Start)
     const services = flags.services
     const environment = flags.environment
     const dryRun = flags['dry-run']
@@ -23,7 +23,7 @@ export default class ReStart extends DockerComposeCommand {
     try {
       this
         .dockerCompose(dryRun)
-        .restart(services, environment)
+        .start({}, services, environment)
     } catch (e) {
       this.error(e.message, e)
     }
