+++ test/commands/exec.test.ts
@@ -6,7 +6,7 @@ describe('exec', () => {
   test
     .env(env)
     .stdout()
-    .command(['exec', 'bin/bash', '--service', 'api', '--dry-run'])
+    .command(['service:exec', 'bin/bash', '--service', 'api', '--dry-run'])
     .it('invokes exec with known service', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('exec', mainConfig.compose.defaultEnvironment, ['api', 'bin/bash']))
     })
