+++ test/commands/logs.test.ts
@@ -6,7 +6,7 @@ describe('logs', () => {
   test
     .env(env)
     .stdout()
-    .command(['logs', '--services', 'api', '--dry-run'])
+    .command(['service:logs', '--services', 'api', '--dry-run'])
     .it('invokes logs with known service', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('logs', mainConfig.compose.defaultEnvironment, ['api']))
     })
@@ -14,7 +14,7 @@ describe('logs', () => {
   test
     .env(env)
     .stdout()
-    .command(['logs', '--services', 'api', '--follow', '--dry-run'])
+    .command(['service:logs', '--services', 'api', '--follow', '--dry-run'])
     .it('invokes logs with follow flag', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('logs', mainConfig.compose.defaultEnvironment, ['--follow', 'api']))
     })
@@ -22,7 +22,7 @@ describe('logs', () => {
   test
     .env(env)
     .stdout()
-    .command(['logs', '--services', 'api', '--timestamps', '--dry-run'])
+    .command(['service:logs', '--services', 'api', '--timestamps', '--dry-run'])
     .it('invokes logs with timestamps flag', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('logs', mainConfig.compose.defaultEnvironment, ['--timestamps', 'api']))
     })
