+++ test/commands/run.test.ts
@@ -6,7 +6,7 @@ describe('run', () => {
   test
     .env(env)
     .stdout()
-    .command(['run', 'bin/bash', '--service', 'api', '--dry-run'])
+    .command(['service:run', 'bin/bash', '--service', 'api', '--dry-run'])
     .it('runs run "bin/bash" --service api', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('run', mainConfig.compose.defaultEnvironment, ['api', 'bin/bash']))
     })
