+++ test/commands/status.test.ts
@@ -6,7 +6,7 @@ describe('status', () => {
   test
     .env(env)
     .stdout()
-    .command(['status', '--services', 'api', '--dry-run'])
+    .command(['service:status', '--services', 'api', '--dry-run'])
     .it('runs status --service api', ctx => {
       expect(ctx.stdout).to.contain(expectedStdOutForCmd('status', mainConfig.compose.defaultEnvironment, ['api']))
     })
